backend/requirements.txt
# Minimal Phase 1 Requirements

# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0

# Database - MongoDB with Beanie ODM
motor==3.6.0
beanie==1.26.0
pymongo==4.9.0

# LLM Providers
openai

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]>=3.4.8
python-multipart==0.0.12

# Environment
python-dotenv==1.1.0

# HTTP Client (für spätere APIs)
httpx==0.28.0

# Dev Tools
pytest==8.3.0
pytest-asyncio==0.25.0
